# docs/openapi.yaml
openapi: 3.0.3
info:
  title: e-SPPD API
  description: |
    API untuk Sistem Elektronik Surat Perintah Perjalanan Dinas (e-SPPD).

    ## Authentication
    API menggunakan Laravel Sanctum untuk autentikasi. Dapatkan token dengan endpoint `/auth/login`.

    ## Rate Limiting
    API dibatasi 60 request per menit per user.
  version: 1.0.0
  contact:
    email: admin@esppd.example.com
servers:
  - url: http://127.0.0.1:8000/api
    description: Development Server
  - url: https://esppd.example.com/api
    description: Production Server
...
paths:
...
  /sppd:
    get:
      tags: [SPPD]
      summary: List SPPD
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, submitted, approved, rejected, completed]
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
        - name: per_page
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List SPPD
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SppdListResponse"

    post:
      tags: [SPPD]
      summary: Create SPPD
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SppdCreateRequest"
      responses:
        "201":
          description: SPPD created
        "422":
          description: Validation error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    SppdListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Sppd"
        meta:
          type: object
          properties:
            current_page:
              type: integer
            total:
              type: integer
    Sppd:
      type: object
      properties:
        id:
          type: string
          format: uuid
        spt_number:
          type: string
        spd_number:
          type: string
        destination:
          type: string
        purpose:
          type: string
        departure_date:
          type: string
          format: date
        return_date:
          type: string
          format: date
        status:
          type: string
          enum: [draft, submitted, approved, rejected, completed]
    SppdCreateRequest:
      type: object
      required:
        - employee_id
        - destination
        - purpose
        - departure_date
        - return_date
        - transport_type
        - budget_id
      properties:
        employee_id:
          type: string
          format: uuid
        destination:
          type: string
        purpose:
          type: string
        departure_date:
          type: string
          format: date
        return_date:
          type: string
          format: date
        transport_type:
          type: string
          enum: [pesawat, kereta, bus, mobil_dinas, kapal]
        budget_id:
          type: string
          format: uuid
        invitation_number:
          type: string
...